added logout protection

+++ frontend/js/components/userCard/userCard.js
@@ -51,6 +51,7 @@ function settingsListener() {
 async function logoutUser() {
     console.log('logout!')
     const logoutResponse = await fetchAuth('POST', 'logout/')
+    if (!logoutResponse) { return }
     if (logoutResponse.status == 200) {
         sessionStorage.clear()
         navigateTo('/')
